+++ BLService/BLService.csproj
@@ -56,9 +56,9 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="IAPIService.cs" />
+    <Compile Include="IBLService.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="APIService.cs" />
+    <Compile Include="BLSevice.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
