+++ BLService/IBLService.cs
@@ -1,0 +1,41 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Runtime.Serialization;
+using System.ServiceModel;
+using System.Text;
+using HtmlAgilityPack;
+using System.Collections;
+
+namespace BLService
+{
+    [ServiceContract]
+    public interface IBLService
+    {
+        [OperationContract]
+        List<Product> GetPrices(byte[] image);
+
+        [OperationContract]
+        List<Shop> GetShops();
+    }
+
+    [DataContract]
+    public class Product
+    {
+        [DataMember]
+        public string Name { get; set; }
+
+        [DataMember]
+        public string Price { get; set; }
+    }
+
+    [DataContract]
+    public class Shop
+    {
+        [DataMember]
+        public long Id { get; set; }
+
+        [DataMember]
+        public string Name { get; set; }
+    }
+}
