--- ConsoleClientTest/tessdata/eng.cube.fold
@@ -1,12 +1,0 @@
-0oO
-lI1
-cC
-kK
-pP
-sS
-uU
-vV
-wW
-xX
-yY
-zZ
