--- ConsoleClientTest/tessdata/eng.cube.lm
@@ -1,7 +1,0 @@
-LeadPunc="({[`'
-TrailPunc=}:;-]!?`,.)"'
-NumLeadPunc=#({[@$
-NumTrailPunc=}):;].,%
-Operators=*+-/.:,()[]
-Digits=0123456789
-Alphas=abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ
