--- ConsoleClientTest/tessdata/eng.cube.params
@@ -1,14 +1,0 @@
-RecoWgt=1.0
-SizeWgt=0.2435
-OODWgt=0.0214
-NumWgt=0.036
-CharBigramsWgt=0.1567
-MaxSegPerChar=8
-BeamWidth=10
-ConvGridSize=48
-WordUnigramsWgt=0.01
-MaxWordAspectRatio=20.0000
-MinSpaceHeightRatio=0.5000
-MaxSpaceHeightRatio=0.6000
-HistWindWid=2
-MinConCompSize=0
